backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
slowapi==0.1.9
pymupdf==1.24.5
python-docx==1.1.2
scipy==1.13.1
pydub==0.25.1
TTS==0.22.0
torch==2.3.1
torchaudio==2.3.1
python-magic==0.4.27
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
